Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.1
httpx==0.27.0
pydantic==2.9.0
python-dotenv==1.0.1